requirements.txt
 Django>=2.0
 psycopg2
 djangorestframework
